backend/requirements/base.txt
django==4.2.4

djangorestframework==3.14.0
django-cors-headers==4.2.0
drf-spectacular==0.26.4
djoser==2.2.0

pyotp==2.9.0
celery[redis]==5.3.4

psycopg2==2.9.7
gunicorn==21.2.0